move MetadataAnnotationsCell into it's own file
